param/store: add parameter evaluation engine
Add a simple engine for parameter evaluation. When a parameter set is
loaded add an entry to DEPENDENCIES dict for each parameter. The entry
contains a list of parameter descriptors for paramters that depend on
given parameter, ex. if parameter C, depends on A and B, then parameter
A will have C in its list and similarly for B. The list is used for
reverse lookup, so that when a paramter A is changed, we can identify
the paramters that require an update.

Change-Id: I55d12acf7c3d3e83a1f0053aaba504d255ca4cb1