sysparams: make shutter_us settable
Allow for setting shutter_us via web API. Ideally we should only set a
single parameter (ex. shutter_us), and have the other automatically
calculated by the camera. Since we don't have the functionality of
updating of all affected camera parameters when a paramter is set,
assume that we allow for modification of both shutter_deg and
shutter_us.

Change-Id: I3b3387617c7de1faec1ab3cb3146e84260abd118